feat: implementacao de alunosDTO

+++ src/main/java/br/com/crud/personalcontroller/database/model/Alunos.java
@@ -2,15 +2,13 @@ package br.com.crud.personalcontroller.database.model;
 
 import jakarta.persistence.*;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Table(name = "alunos")
 public class Alunos {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "NOME", nullable = false)
     private String nome;
